Move Linkage to a more private location
