+++ link-grammar/analyze-linkage.c
@@ -11,11 +11,12 @@
 /*                                                                       */
 /*************************************************************************/
 
-
+#include <ctype.h>
 #include <string.h>
 #include "analyze-linkage.h"
 #include "string-set.h"
-#include "structures.h"
+#include "linkage.h"
+#include "structures.h" // Needed for connector_get_string
 
 /**
  * This returns a string that is the the GCD of the two given strings.
