+++ link-grammar/api-structures.h
@@ -109,15 +109,6 @@ struct Connector_set_s
 	unsigned int table_size;
 };
 
-struct Link_s
-{
-	size_t lw;              /* Offset into Linkage->word NOT Sentence->word */
-	size_t rw;              /* Offset into Linkage->word NOT Sentence->word */
-	Connector * lc;
-	Connector * rc;
-	const char * link_name; /* Spelling of full link name */
-};
-
 struct Parse_info_struct
 {
 	unsigned int   x_table_size;
@@ -178,71 +169,6 @@ struct Sentence_s
 #ifdef USE_SAT_SOLVER
 	void *hook;                 /* Hook for the SAT solver */
 #endif /* USE_SAT_SOLVER */
-};
-
-/*********************************************************
- *
- * Linkages
- *
- **********************************************************/
-
-/**
- * This summarizes the linkage status.
- */
-struct Linkage_info_struct
-{
-	int index;            /* Index into the parse_set */
-	bool discarded;
-	short N_violations;
-	short unused_word_cost;
-	short link_cost;
-
-	double disjunct_cost;
-	double corpus_cost;
-	const char *pp_violation_msg;
-};
-
-/**
- * num_links:
- *   The number of links in the current linkage.  Computed by
- *   extract_linkage().
- *
- * chosen_disjuncts[]
- *   This is an array pointers to disjuncts, one for each word, that is
- *   computed by extract_links().  It represents the chosen disjuncts
- *   for the current linkage.  It is used to compute the cost of the
- *   linkage, and also by compute_chosen_words() to compute the
- *   chosen_words[].
- *
- * link_array[]
- *   This is an array of links.  These links define the current linkage.
- *   It is computed by extract_links().  It is used by analyze_linkage().
- */
-struct Linkage_s
-{
-	WordIdx         num_words;    /* Number of (tokenized) words */
-	bool            is_sent_long; /* num_words >= twopass_length */
-	const char *  * word;         /* Array of word spellings */
-
-	size_t          num_links;    /* Number of links in array */
-	Link *          link_array;   /* Array of links */
-	size_t          lasz;         /* Alloc'ed length of link_array */
-
-	Disjunct **     chosen_disjuncts; /* Disjuncts used, one per word */
-	size_t          cdsz;         /* Alloc'ed length of chosen_disjuncts */
-	char **         disjunct_list_str; /* Stringified version of above */
-#ifdef USE_CORPUS
-	Sense **        sense_list;   /* Word senses, inferred from disjuncts */
-#endif
-
-	Gword **wg_path;              /* Linkage Wordgraph path */
-	Gword **wg_path_display;      /* ... for !morphology=0. Experimental. */
-	//size_t *wg_path_index;      /* Displayed-word indices in wg_path (FIXME?)*/
-
-	Linkage_info    lifo;         /* Parse_set index and cost information */
-	PP_info *       pp_info;      /* PP domain info, one for each link */
-
-	Sentence        sent;         /* Used for common linkage data */
 };
 
 // XXX FIXME
