+++ link-grammar/post-process/constituents.c
@@ -16,7 +16,7 @@
 
 #include "api-structures.h"
 #include "error.h"
-#include "externs.h"
+#include "linkage.h"
 #include "post-process/post-process.h"
 #include "post-process/pp-structures.h"
 #include "print/print-util.h"
